+++ data/actions/DefendAction.js
@@ -1,0 +1,9 @@
+class Defend {
+    constructor(monster) {
+        this._monster = monster;
+    }
+
+    execute(monster) {
+        monster.activateDefense();
+    }
+}
